+++ helpers/date/getTime.js
@@ -1,0 +1,8 @@
+let time = new Date()
+let hours, minutes
+
+export const getHoursAndMinutesNow = () => {
+    hours = time.getHours()
+    minutes = time.getMinutes()
+    return `${hours}:${minutes}`
+}
