+++ helpers/getRandomNumber.js
@@ -1,0 +1,1 @@
+export const getRandomNumber = (num) => Math.ceil(Math.random() * num)
