+++ helpers/helth_points/elementHP.js
@@ -1,0 +1,3 @@
+export function elementHP() {
+    return document.querySelector(`.player${this.player} .life`)
+}
