+++ helpers/helth_points/renderHP.js
@@ -1,0 +1,3 @@
+export function renderHP() {
+    this.elementHP().style.width = this.hp + '%'
+}
