+++ helpers/logs_and_banners/commentInChat.js
@@ -1,0 +1,1 @@
+export const commentInChat = (time,text, damage, hp) => `<p>${time} ${text} ${damage} ${hp}</p>`
