+++ index.html
@@ -1,5 +1,5 @@
 <!doctype html>
-<html lang="en">
+<html lang="ru">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport"
@@ -57,6 +57,6 @@
         <div class="wall-left"></div>
         <div class="wall-right"></div>
     </div>
-    <script src="main.js"></script>
+    <script src="main.js" type="module"></script>
 </body>
 </html>
