+++ store/body_parts_to_attack.js
@@ -1,0 +1,1 @@
+export const ATTACK = ['head', 'body', 'foot'];
