+++ store/hit_points.js
@@ -1,0 +1,5 @@
+export const HIT = {
+    head: 80,
+    body: 25,
+    foot: 20,
+}
